changed type of 'population field' from double to int

+++ src/main/java/by/itclass/model/entities/Country.java
@@ -10,5 +10,5 @@ public class Country {
     private String name;
     private String capital;
     private String president;
-    private double population;
+    private int population;
 }
